+++ src/gameObjects/index.ts
@@ -1,2 +1,3 @@
 export * from './containerUtils';
 export * from './isSomeGameObject';
+export * from './getArrayBounds';