+++ src/layout/arrayAlign/types.ts
@@ -1,5 +1,4 @@
 import { AlignObject } from "../Align/Align";
-import { IBoundable } from "../types";
 
-export type ILayoutTarget = AlignObject & IBoundable;
+export type ILayoutTarget = AlignObject;
 export type ArrayAlignCallback = <T extends ILayoutTarget>(objects: T[], ...args: number[]) => T[]